Windows build, release builds and GitHub Actions (#3)
This is basically all fiddly configuration, so I didn't even try to record this.
I'll probably just briefly summarize this in the next video.

I added a `npm start` script in ui to make starting the app in dev mode easier. I also added
typescript as a formal dependency and eslint, which can be run via `npm test`. The settings
are very lax, but these will hopefully help me catch silly mistakes.

To my surprise, Oxygen basically just worked on Windows. The only change was enabling
the `bundled` flag of rusqlite. I could have also provided a compiled version of sqlite, but
I figured it was just easier to ask rusqlite to provide a copy.

To build a production app, I used https://www.electronforge.io. I'm not worrying about
installers, code signing, or updates yet, so I only enabled the zip "maker". In pull requests,
it tests and builds the app on all 3 platforms, and uploads it as an asset. Assets are a bit
buried, but if you go to the workflow logs and then click "Summary", you can see and
download them. You need to be downloaded into GitHub. The zip files are double-zipped
due to a limitation.

Because I don't have code signing, you have to go through some silly steps to actually run the app:
 - On Windows, SmartScreen will block the app the first time you try to run it. To continue, click "More info", then "Run anyway". Windows might also determine Electron to be a virus and block you entirely, apparently.
 - On macOS, to be allowed to open the app the first time, right-click and choose "Open"

When I merge this PR, this should create a draft release with the appropriate assets.